+++ app/(auth)/create-account/_components/personal-form-section.tsx
@@ -7,6 +7,7 @@ import { ArrowUpRight } from 'lucide-react'
 import OvationService from '@/services/ovation.service'
 import { FormField } from '@/components/customs/custom-form'
 import colors from '@/lib/colors'
+import GoogleAuth from '../../../../components/google-auth'
 
 export default function PersonalInfoForm({
   setPage,
@@ -14,7 +15,6 @@ export default function PersonalInfoForm({
   setPage: (page: number) => void
 }) {
   const form = useFormContext()
-
   const { mutate: checkUsername, isPending } = useMutation({
     mutationFn: OvationService.checkUsername,
     onSuccess: () => {
@@ -42,6 +42,12 @@ export default function PersonalInfoForm({
       checkUsername(username)
     }
   }
+  const signUpGoogle = ({ googleObject }: any) => {
+    form.setValue('personalInfo.displayName', googleObject.name)
+    form.setValue('personalInfo.email', googleObject.email)
+    form.setValue('type', 'Google')
+    setPage(2)
+  }
 
   return (
     <form onSubmit={handleContinue} className="flex flex-col gap-7">
@@ -59,6 +65,7 @@ export default function PersonalInfoForm({
           <p className="text-[10px] font-medium text-[#C1C0C6]">OR</p>
           <span className="w-[46%] h-[1px] border-[#C1C0C6] border-b-0 border-[1px]" />
         </div> */}
+      <GoogleAuth func={signUpGoogle} />
       <FormField
         name="personalInfo.displayName"
         form={form}
